:recycle: useMenuPosition 으로 position 을 다루는 로직 분리
